Run the full test suite in WebAssembly CI
Add a test_wasm_suite job that runs the whole pytest suite in a Pyodide
virtual environment (pyodide-build + node), with example files seeded
natively into the pooch cache via DFS_DATA_DIR so no network access is
needed inside the runtime.

Tests that spawn threads or processes can never pass on WebAssembly, so
they are tagged with a new 'concurrency' marker and deselected along
with 'network' tests in the wasm job only.

Also fix three portability bugs the wasm (32-bit) run exposed:
- _float_array_to_ns overflowed for integer input arrays on platforms
  where the default integer is 32 bits.
- The ProdML writer stored Count/StartIndex/NumberOfLoci/StartLocusIndex
  HDF5 attrs as platform-native ints (int32 on wasm32) instead of int64.
- dascore.utils.progress.track crashed on WebAssembly because rich
  progress needs a refresh thread; it now falls back to plain iteration.

+++ dascore/io/prodml/utils.py
@@ -359,11 +359,11 @@ def _get_acquisition_attrs(
         "MeasurementStartTime": start_time,
         "MinimumFrequency": 0.0,
         "MinimumFrequency.uom": encode_h5_strings("Hz")[0],
-        "NumberOfLoci": num_loci,
+        "NumberOfLoci": np.int64(num_loci),
         "ServiceCompanyName": encode_h5_strings(strings["ServiceCompanyName"])[0],
         "SpatialSamplingInterval": distance_step,
         "SpatialSamplingInterval.uom": encode_h5_strings("m")[0],
-        "StartLocusIndex": start_locus,
+        "StartLocusIndex": np.int64(start_locus),
         "TriggeredMeasurement": False,
         "schemaVersion": encode_h5_strings("2.1")[0],
         "uuid": encode_h5_strings(str(uuid4()))[0],
@@ -385,30 +385,30 @@ def _get_array_attrs(
     """Return attributes for the raw data and time arrays."""
     num_loci = len(patch.get_coord("distance"))
     raw_attrs = {
-        "NumberOfLoci": num_loci,
+        "NumberOfLoci": np.int64(num_loci),
         "OutputDataRate": output_rate,
         "OutputDataRate.uom": encode_h5_strings("Hz")[0],
         "RawDataUnit": encode_h5_strings(strings["RawDataUnit"])[0],
-        "StartLocusIndex": start_locus,
+        "StartLocusIndex": np.int64(start_locus),
         "uuid": encode_h5_strings(str(uuid4()))[0],
     }
     if patch.attrs.data_type:
         raw_attrs["RawDescription"] = encode_h5_strings(patch.attrs.data_type)[0]
     data_attrs = {
-        "Count": patch.data.size,
+        "Count": np.int64(patch.data.size),
         "Dimensions": encode_h5_strings(
             ["locus" if dim == "distance" else dim for dim in patch.dims]
         ),
         "PartEndTime": end_time,
         "PartStartTime": start_time,
-        "StartIndex": 0,
+        "StartIndex": np.int64(0),
     }
     time_attrs = {
-        "Count": len(times),
+        "Count": np.int64(len(times)),
         "EndTime": end_time,
         "PartEndTime": end_time,
         "PartStartTime": start_time,
-        "StartIndex": 0,
+        "StartIndex": np.int64(0),
         "StartTime": start_time,
         "Uom": encode_h5_strings("us")[0],
     }
